FSM now inherits from comb, move reg and register inference methods to comb
